### 2025-03-28 #695 @marcboschmatas
partialling.out: Residuals from partial regressions
Submitting Author Name: Marc Bosch-Matas
Submitting Author Github Handle: @marcboschmatas
Repository:  https://github.com/marcboschmatas/partialling.out
Submission type: Pre-submission

---

```
Package: partialling.out
Title: Residuals from partial regressions
Version: 0.0.1
    person("Marc", "Bosch-Matas", , "mboschmatas@gmail.com", role = c("aut", "cre"),
Description: Creates a data frame with the residuals of partial regressions of the main explanatory variable and the variable of interest. This method follows the Frisch-Waugh-Lovell theorem, as explained in Lovell (2008) <doi:10.3200/JECE.39.1.88-91>.
    tinytest,
    palmerpenguins,
    tinytable,
    tinyplot,
    fwlplot
Roxygen:list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    glue,
    lifecycle,
    fixest,
    lfe,
    stats
```

The package residualises the outcome and main explanatory variables of an already existing model. This allows for easier model visualisations (e.g. by using scatterplots) or as a basis for further analysis. I doubt whether to include it in Regression and Supervised Learning, but the only thing it does is to residualise these variables, implementing already existing algorithms under the hood.

Yes

Econometricians who intend to visualise models via scatter plots or similar tools. Transform multivariate into bivariate models for simpler visualisation or analysis.

`fwlplot` provides a similar case but it just returns the scatterplot of the two residualised variables instead of a data.frame with said residuals.
