+++ README.Rmd
@@ -83,9 +83,11 @@ tt(head(res)) |>
 The Frisch-Waugh-Lovell theorem states that for a linear model
 
 $$ Y = X_1 \beta_1 + X_2\beta_2 + u $$
+
 The coefficient $\beta_2$ will be equivalent to that in the regression
 
 $$ M_{X_1} Y = M_{X_1}X_2\beta_2 + M_{X_1}u $$
+
 Where $M_{X_1}Y$ are the residuals of the model
 
 $$ Y = X_1 \beta_1 + u $$
