### 2025-03-30 #695 @marcboschmatas
Hi, thanks for the checks. I'm not sure I understand
"✖️ Statistical standards should be documented in most package files, yet are mostly only documented in one file.". How should I proceed here?
