### 2025-03-30 38b350a
cleaning and explaing srr standards, bumped version
