+++ NEWS.md
@@ -3,3 +3,9 @@
 ## partialling.out 0.0.1
 
 First version
+
+
+## partialling.out 0.0.2
+
+Updated test structure and documentation to comply with rOpensci standards
+
