### 2025-03-30 bfa1d8e
cleaning srr standards
