### 2025-03-30 #695 @maurolepore
@marcboschmatas thanks so much for sharing your work with rOpenSci. As this is a stats pre-submission I would expect @mpadge to somment here soon.
