### 2025-03-31 #695 @mpadge
Thanks @marcboschmatas, I just opened https://github.com/marcboschmatas/partialling.out/pull/1 to fix the issue. `roxygen2` works by identifying "blocks" which need to be some kind of valid R thing. Your `srr-stats` file had comments without anything else, so there was nothing for R to identify. The PR just adds `NULL` to the end, which then allows roxygen2 to identify that as a block. Those `srrstats` tags are then captured appropriately.

That then indicates the you've complied with 57 / 102 applicable standards, or 56%. While our minimal requirement is at least 50%, it does seem from a preliminary scan that you should be able to comply with a few more, and increase this proportion somewhat. Examples include standards for input in G2.11, G2.12, G2.13, G2.14, as well as EA5.3, EA6.0, and possibly EA2.0 as an extension. Just adding a few more robust input checking routines should increase compliance a fair bit. Let us know what you think.
