### 2025-03-31 #695 @marcboschmatas
Hello @mpadge . Sounds fair. I'll be giving it a go and let you know if I need some help.
I can imagine some sort of check for G2.11, G2.12, and G.13. I thought G2.14 would be handled by passing optional arguments in `...` to `lm()`, `feols()`, or `felm()`, but I could make it explicit to remove observations with NA instead of allowing the underlying functions to to their thing.
I thought I had incorporated EA6.0 enough, but I'll give it a look to see what else I can incorporate.
Regarding EA5.3, I guess I could test that any summary statistics computed by user (via `summary()` or whatever) work as expected.
Regarding EA2.0 I'm a bit more at a loss, but I'll give it some thought.

Does that sound OK?
