### 2025-04-03 7f794d0
partialling_out returns weights if passed, new tests, updated vignettes
