+++ R/partialling_out.R
@@ -12,11 +12,15 @@
 #' Will accept lm, felm (lfe package), and feols (fixest package) objects
 #' @title partialling_out: partialling out variable of interest and main
 #' @param model object for which we want to residualise variables
-#' @param data data.frame used in the original model
+#' @param data data.frame used in the original model. Using different data will
+#' return unexpected results or an error.
 #' @param weights a numeric vector for weighting the partial models
 #' @param both if `TRUE` will residualise both the variable of interest and the
 #' first explanatory variable in the model. If `FALSE`, only the latter.
 #' Set to `TRUE` by default
+#' @param na.rm if `TRUE` will remove observations with NA before any models are
+#' run. If `FALSE`, the underlying `lm`, `feols`, or `felm` will remove NA
+#' values but errors may arise if weights are used.
 #' @param ... Any other lm, feols, or felm parameters that will be passed to the
 #' partial regressions
 #' @returns a data.frame with the (residualised) variable of interest and
@@ -33,7 +37,9 @@
 #' @srrstats {G2.1} *Implement assertions on types of inputs (see the initial point on nomenclature above).*
 #' @srrstats {G2.1a} *Provide explicit secondary documentation of expectations on data types of all vector inputs.*
 #' @srrstats {G2.14} *Where possible, all functions should provide options for users to specify how to handle missing (`NA`) data, with options minimally including:*
+#' @srrstats {G2.13} *Statistical Software should implement appropriate checks for missing data as part of initial pre-processing prior to passing data to analytic algorithms.*
 #' @srrstats {G2.14a} *error on missing data*
+#' @srrstats {G2.14b} *ignore missing data with default warnings or messages issued*
 #' @srrstats {G2.15} *Functions should never assume non-missingness, and should never pass data with potential missing values to any base routines with default `na.rm = FALSE`-type parameters (such as [`mean()`](https://stat.ethz.ch/R-manual/R-devel/library/base/html/mean.html), [`sd()`](https://stat.ethz.ch/R-manual/R-devel/library/stats/html/sd.html) or [`cor()`](https://stat.ethz.ch/R-manual/R-devel/library/stats/html/cor.html)).*
 #' @srrstats {G2.16} *All functions should also provide options to handle undefined values (e.g., `NaN`, `Inf` and `-Inf`), including potentially ignoring or removing such values.*
 #' @srrstats {G3.0} *Statistical software should never compare floating point numbers for equality. All numeric equality comparisons should either ensure that they are made between integers, or use appropriate tolerances for approximate equality.*
@@ -57,6 +63,7 @@
 #' @srrstats {G5.9} **Noise susceptibility tests** *Packages should test for expected stochastic behaviour, such as through the following conditions:*
 #' @srrstats {G5.9a} *Adding trivial noise (for example, at the scale of `.Machine$double.eps`) to data does not meaningfully change results*
 #' @srrstats {G5.9b} *Running under different random seeds or initial conditions does not meaningfully change results*
+#' @srrstats {G2.10} *Software should ensure that extraction or filtering of single columns from tabular inputs should not presume any particular default behaviour, and should ensure all column-extraction operations behave consistently regardless of the class of tabular data used as input.*
 #' @srrstats {EA1.0} *Identify one or more target audiences for whom the software is intended*
[…]
