### 2025-04-03 #695 @marcboschmatas
Hi, I've gone a over all these conditions. Plus, I've changed the function so that whenever weights are used, it will return the weights vector as a column of the result df in case the user wants to build a model using the residualised variables.

G2.11: I'm not sure how to go over this one, I'm afraid of getting unpredictable results, so I'd rather go with whatever lm, felm, or feols go with. However, I'd be happy to change it if needed.
G2.12: list columns are not accepted by the underlying lm, felm, or feols models, and since the user is expected to use the same df as the original model (made it a bit clearer in the documentation), I think it's better to return an error if any list-columns are passed to the function.
G2.13 & G2.14: added a `na.rm` argument that will remove any rows with a NA value of the columns used in the model, and the same indexes for the weights vector, if used. It will throw out a warning that results might be unexpected or that the underlying functions can throw an error.
EA5.3: added a test to make sure summaries of the columns returned by the
EA2.0: I don't really see how to add an index column, there are no joins, and there is just some column selection, filtering and column binding at the end. What was your idea here? I'm open to changing it.
