### 2025-04-04 #695 @mpadge
Thanks @marcboschmatas. Note the final line of the bot response there - now can you please try to move your `srrstats` tags out of the `srr-stats-standards.R` file and place them in your actual where each is actually addressed. This will often require repeated versions of the same standard.

Also could you please think a bit more about the standards still documented as `srrstatsNA`, as many of these clearly already apply. This reprex demonstrates that G2.7, G2.11, and G2.12 are already applicable, and compliance would just require proper testing and issuing of additional messages where appropriate:
``` r
library(partialling.out)
n <- 100L
df <- tsibble::tsibble(
    date = as.Date("2017-01-01") + seq_len(n),
    dv = rnorm(n),
    iv1 = rnorm(n),
    iv2 = sample(letters, size = n, replace = TRUE),
    iv3 = I(as.list(sample(letters, size = n, replace = TRUE)))
)
#> Registered S3 method overwritten by 'tsibble':
#>   method               from
#>   as_tibble.grouped_df dplyr
#> Using `date` as index variable.
model <- feols(dv ~ iv1 | iv2 + iv3, data = df)
partial_df <- partialling_out(model, df, both = TRUE)
```

The input there is a non-standard `data.frame` with a list column, and everything works fine, so G2.7, 2.11, and 2.12 are already satisfied in principle. That also likely means that the intervening standards, G2.8-G2.10, would also apply. Those are just ones that I could briefly see as applicable, and there may be others as well. Could you please have another go at increasing compliance? Thanks!
