### 2025-04-04 #695 @marcboschmatas
Will do! Thanks a lot. I need to make sure I understand what those standards do and how to properly test them. I guess that standards documentation can be placed also in the `/inst/tinytest` folder just fine, right?
