### 2025-04-04 #695 @mpadge
> Will do! Thanks a lot. I need to make sure I understand what those standards do and how to properly test them. I guess that standards documentation can be placed also in the `/inst/tinytest` folder just fine, right?

Yep, although i checked and noticed a slight glitch, so you'll need to ensure you've got the latest version of `srr` installed.
