### 2025-04-05 7e1bdc0
added some warnings and tests, restructured srr standards
