+++ inst/tinytest/test_partialling.out.R
@@ -5,10 +5,34 @@ library(fixest)
 library(lfe)
 library(tinytest)
 library(partialling.out)
+library(tsibble)
+library(units)
 
 penguins <- penguins
 set.seed(1234)
-# Expect error if no data is provided ----
+
+# general standards for tests
+
+# nolint start
+#' @srrstats {G5.0} *Where applicable or practicable, tests should use standard data sets with known properties (for example, the [NIST Standard Reference Datasets](https://www.itl.nist.gov/div898/strd/), or data sets provided by other widely-used R packages).*
+#' @srrstats {G5.1} *Data sets created within, and used to test, a package should be exported (or otherwise made generally available) so that users can confirm tests and run examples.*
+#' @srrstats {G5.3} *For functions which are expected to return objects containing no missing (`NA`) or undefined (`NaN`, `Inf`) values, the absence of any such values in return objects should be explicitly tested.*
+NULL
+# NOLINT END
+
+
+# Test for errors warnings etc ----
+
+# nolint start
+#' @srrstats {G5.2} *Appropriate error and warning behaviour of all functions should be explicitly demonstrated through tests. In particular,*
+#' @srrstats {G5.2a} *Every message produced within R code by `stop()`, `warning()`, `message()`, or equivalent should be unique*
+#' @srrstats {G5.2b} *Explicit tests should demonstrate conditions which trigger every one of those messages, and should compare the result with expected values.*
+NULL
+
+# nolint end
+
+
+## Expect error if no data is provided ----
 
 expect_error({
               mod <- lm(bill_length_mm ~  bill_depth_mm + species,
@@ -28,7 +52,7 @@ expect_error({
   partialling_out(mod)})
 
 
-# expect error if data is not a data.frame ----
+## expect error if data is not a data.frame ----
 
 expect_error({
               mod <- lm(bill_length_mm ~  bill_depth_mm + species,
@@ -48,7 +72,7 @@ expect_error({
   partialling_out(mod, data = c(1,2,3))})
 
 
-# expect error if weights are non numeric ----
+## expect error if weights are non numeric ----
 
 expect_error({
   mod <- lm(bill_length_mm ~  bill_depth_mm + species, data = penguins)
@@ -65,7 +89,7 @@ expect_error({
   partialling_out(mod, data = penguins, weights = rep("a", nrow(penguins)))})
 
 
-# expect error if weights are of the wrong length ----
+## expect error if weights are of the wrong length ----
 
 expect_error({
   mod <- lm(bill_length_mm ~  bill_depth_mm + species, data = penguins)
@@ -81,7 +105,7 @@ expect_error({
   partialling_out(mod, data = penguins, weights = c(1,2,3))})
 
 
-# expect error if only one covariate or no fixed effects ----
+## expect error if only one covariate or no fixed effects ----
 
 expect_error({
   mod <- lm(bill_length_mm ~  bill_depth_mm, data = penguins)
@@ -99,7 +123,7 @@ expect_error({
 
 
 
-# expect a warning if partial models are weighted but original model is not ----
+## expect a warning if partial models are weighted but original model is not ----
 
 
 expect_warning({
@@ -119,7 +143,7 @@ expect_warning({
 
 
 
-# expect a warning if original model is weighted but partialling out is not ----
+## expect a warning if original model is weighted but not partialling_out ----
 
 
 expect_warning({
@@ -142,8 +166,184 @@ expect_warning({
   partialling_out(mod, data = penguins[!is.na(penguins$body_mass_g), ])})
 
 
+
[…]
