+++ R/partialling.out-package.R
@@ -2,8 +2,21 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+
+# nolint start
+#' @srrstats {G1.0} *Statistical Software should list at least one primary reference from published academic literature.*
+#' @srrstats {G1.1} *Statistical Software should document whether the algorithm(s) it implements are:* - *The first implementation of a novel algorithm*; or - *The first implementation within **R** of an algorithm which has previously been implemented in other languages or contexts*; or - *An improvement on other implementations of similar algorithms in **R***.
+#' @srrstats {G1.2} *Statistical Software should include a* Life Cycle Statement *describing current and anticipated future states of development.*
+#' @srrstats {G1.3} *All statistical terminology should be clarified and unambiguously defined.*
+#' @srrstats {G1.4} *Software should use [`roxygen2`](https://roxygen2.r-lib.org/) to document all functions.*
+#' @srrstats {G1.4a} *All internal (non-exported) functions should also be documented in standard [`roxygen2`](https://roxygen2.r-lib.org/) format, along with a final `@noRd` tag to suppress automatic generation of `.Rd` files.*
+#' @srrstats {G1.5} *Software should include all code necessary to reproduce results which form the basis of performance claims made in associated publications.*
+#' @srrstats {G1.6} *Software should include code necessary to compare performance claims with alternative implementations in other R packages.*
+#' @srrstats {EA1.0} *Identify one or more target audiences for whom the software is intended*
 #' @import rlang
 #' @importFrom glue glue
 #' @importFrom lifecycle deprecated
 ## usethis namespace: end
 NULL
+
+# nolint end
