+++ DESCRIPTION
@@ -14,7 +14,9 @@ Suggests:
     palmerpenguins,
     tinytable,
     tinyplot,
-    fwlplot
+    fwlplot,
+    tsibble,
+    units
 Config/testthat/edition: 3
 Encoding: UTF-8
 Roxygen:list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))

[…]
