### 2025-04-05 df3c5a3
refactored main function, version bump
