+++ NEWS.md
@@ -12,3 +12,8 @@ Updated test structure and documentation to comply with rOpensci standards
 ## partialling.out 0.1.0
 
 Modified the function so that it returns a weights vector as a column of the result df
+
+
+## partialling.out 0.1.1
+
+Refactored main function, taken formula construction into a secondary internal function
