### 2025-04-05 #695 @marcboschmatas
Done that. I've also refactored a bit the main function to put all the working with formulas out of the main function, it was getting a bit too long for my taste.
