### 2025-04-05 #695 @marcboschmatas
This is not what I'm getting with my own srr checks (I've downloaded the latest version), and when I run `pkgcheck::pkgcheck()`, I no longer get the second warning (see results dump (here)[https://gist.github.com/marcboschmatas/caf7d365995507bb3a0093fb678c896a]
I guess it is because the "missing" standards are in `inst/tinytest` only and the bot hasn't caught up in the glitch you mentioned before. I'm happy to clarify further.
