### 2025-04-07 067ea12
minor changes to data class checking
