### 2025-04-07 #695 @mpadge
@marcboschmatas Yeah,  sorry, I should have said that our system only rebuilds once a week. That will happen about 12 hours from now, and it will then include the updated `srr` that should catch your standards in `inst/tinytest`. I've check locally and see everything fine there - 76 complied with and 35 N/A, which is great!
