### 2025-04-08 #695 @mpadge
@marcboschmatas Sorry to inconvenience you, but the `srr` package has just been updated. When you locally run `srr_stats_pre_submit()`, you should now see this:

``` r
✔ All applicable standards [v0.2.0] have been documented in this package (76 complied with; 35 N/A standards)
! Text of standards should document how package complies, not just copy original standards text.
```

And when those changes get incorporated within our deployed version of the bot (early next week), it will report this:

---

:heavy_multiplication_x: Error: Text of standards should document how package complies, not just copy original standards text.

## 'srr' standards compliance:

- Complied with: 67 / 102 = 65.7% (eda: 21 / 34; general: 46 / 68)
- Not complied with: 35 / 102 = 34.3% (eda: 13 / 34; general: 22 / 68)

:heavy_check_mark: This package complies with > 50% of all standards.

---

You need to change the text of the standards so that each one documents how your package actually complies. Please refer to both the [*Stats Dev Guide*](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr) and the [`srr` documentation](https://ropensci-review-tools.github.io/srr/articles/srr-stats.html#4-the-srr-workflow) for details.
