### 2025-04-08 9ae0551
explain srr standards
