### 2025-04-08 10c59fd
explain srr standards
