### 2025-04-15 #695 @marcboschmatas
@mpadge I'm mostly done (I discovered a few bugs with interaction terms that I'll solve this week), but I have a doubt with standard  G5.2b. Does it mean that I should capture the error term and make sure it is what I expect it to be? So far I've just tested every conditional that may give an error or a warning and make sure it does so.
