### 2025-04-15 #695 @mpadge
> [@mpadge](https://github.com/mpadge) I'm mostly done (I discovered a few bugs with interaction terms that I'll solve this week), but I have a doubt with standard G5.2b. Does it mean that I should capture the error term and make sure it is what I expect it to be? So far I've just tested every conditional that may give an error or a warning and make sure it does so.

Thanks for asking @marcboschmatas. Yes, that does mean that your tests should match each error string. That generally means hard-coding repeated versions of all error strings in your test suite.
