### 2025-04-18 121c048
changed test suite, added interactions
