+++ DESCRIPTION
@@ -16,7 +16,8 @@ Suggests:
     tinyplot,
     fwlplot,
     tsibble,
-    units
+    units,
+    purrr
 Config/testthat/edition: 3
 Encoding: UTF-8
 Roxygen:list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))

[…]
