### 2025-04-24 #695 @marcboschmatas
Hi @mpadge, just to let you know, I explained all standards (plus quite a few changes in the test suite), so if you agree, we can move on to the next steps.
