### 2025-04-25 #695 @mpadge
@marcboschmatas Great to see that you were able to comply with so many of the EDA standards. @maurolepore This all looks grand now. I think we can proceed to a full submission.
